type refactoring

+++ 01_MyDevServer/00_TB_IOCP_CHAT/NetLib/NetService.cpp
@@ -256,7 +256,7 @@ void NetService::WokerThread()
 				OnSend(pSession, ioSize);
 				break;
 			default:
-				LOG(ERROR) << "Bad IO TYPE : " << (int)pSession->GetSock();
+				LOG(ERROR) << "Bad IO TYPE : " << pSession->GetSock();
 				break;
 			}
 		}
